Refactor auction-related data types and controllers to use decimal for monetary values and introduce request objects for auction and auction sale operations

+++ LeafBidAPI/App/Domain/Auction/Data/CreateAuctionData.cs
@@ -9,7 +9,7 @@ public record CreateAuctionData(
     string Description,
     DateTime StartDate,
     int Amount,
-    int MinimumPrice,
+    decimal MinimumPrice,
     ClockLocationEnum ClockLocationEnum,
     int AuctioneerId
 );
